Add exceptions for deletion and validation

+++ src/main/java/com/matheusluizago/libraryapi/service/AuthorService.java
@@ -1,7 +1,10 @@
 package com.matheusluizago.libraryapi.service;
 
+import com.matheusluizago.libraryapi.exceptions.OperationNotAllowedException;
 import com.matheusluizago.libraryapi.model.Author;
 import com.matheusluizago.libraryapi.repository.AuthorRepository;
+import com.matheusluizago.libraryapi.repository.BookRepository;
+import com.matheusluizago.libraryapi.validator.AuthorValidator;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,12 +15,17 @@ import java.util.UUID;
 public class AuthorService {
 
     private final AuthorRepository repository;
+    private final AuthorValidator validator;
+    private final BookRepository bookRepository;
 
-    public AuthorService(AuthorRepository repository){
+    public AuthorService(AuthorRepository repository, AuthorValidator validator, BookRepository bookRepository){
         this.repository = repository;
+        this.validator = validator;
+        this.bookRepository = bookRepository;
     }
 
     public Author save(Author author){
+        validator.validate(author);
         return repository.save(author);
     }
 
@@ -25,6 +33,8 @@ public class AuthorService {
         if(author.getId() == null){
             throw new IllegalArgumentException("It is necessarity to have the author already in the database to update it.");
         }
+
+        validator.validate(author);
         repository.save(author);
     }
 
@@ -33,6 +43,9 @@ public class AuthorService {
     }
 
     public void delete(Author author) {
+        if(hasBook(author)){
+            throw new OperationNotAllowedException("It is not allowed to delete an author with registered books!");
+        }
         repository.delete(author);
     }
 
@@ -46,6 +59,10 @@ public class AuthorService {
         } else {
             return repository.findAll();
         }
+    }
+
+    public boolean hasBook(Author author){
+        return bookRepository.existsByAuthor(author);
     }
 
 }
